Refactor / #22 / Refactor : board

+++ services/freeBoard.js
@@ -1,5 +1,6 @@
 const { boardRepo } = require("../repos");
 const model = require("../database/models/freeBoard");
+const bcrypt = require("bcrypt");
 
 const getPosts = async (req, res, next) => {
   try {
@@ -22,13 +23,28 @@ const getPost = async (req, res, next) => {
 
 const setPost = async (req, res, next) => {
   try {
-    const { id, title, content } = req.body;
+    console.log("cponter");
+    const { id } = req.params;
+    const { title, content, password } = req.body;
     // const { id, categoryId, title, content } = req.body;
     // const existingPost = await freeBoardRepos.checkDeletedPost(id);
-    const existingPost = await boardRepo.checkPost(id, model);
+    const existingPost = await boardRepo.findPost(id, model);
+
     if (!existingPost) {
       throw new Error("이미 삭제된 공고입니다.");
     }
+    console.log(req.user.id);
+    console.log(existingPost.User);
+    if (req.user.id !== existingPost.UserId) {
+      throw new Error("글 작성자가 아닙니다.");
+    }
+
+    // 해당 공지 게시글의 글 작성자인지 확인하기 위해 비밀번호를 입력받는다.
+    const isPasswordCorrect = await bcrypt.compare(password, req.user.password);
+    if (!isPasswordCorrect) {
+      throw new Error("비밀번호가 일치하지 않습니다.");
+    }
+
     const updatedPost = await boardRepo.updatePost(id, title, content, model);
     // const updatedPost = await freeBoardRepos.updatePost(
     //   id,
@@ -51,10 +67,25 @@ const setPost = async (req, res, next) => {
 const deletePost = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const deletedPost = await boardRepo.destroyPost(id, model);
-    if (!deletedPost) {
+    const { password } = req.body;
+    const existingPost = await boardRepo.findPost(id, model);
+
+    if (!existingPost) {
       throw new Error("이미 삭제된 공고입니다.");
     }
+
+    if (req.user.id !== existingPost.UserId) {
+      throw new Error("글 작성자가 아닙니다.");
+    }
+
+    // 해당 공지 게시글의 글 작성자인지 확인하기 위해 비밀번호를 입력받는다.
+    const isPasswordCorrect = await bcrypt.compare(password, req.user.password);
+    if (!isPasswordCorrect) {
+      throw new Error("비밀번호가 일치하지 않습니다.");
+    }
+
+    await boardRepo.destroyPost(id, model);
+
     return res.status(200).json({ message: "Posting is deleted" });
   } catch (err) {
     next(err);
@@ -63,12 +94,13 @@ const deletePost = async (req, res, next) => {
 
 const addPost = async (req, res, next) => {
   try {
-    const userId = req.userId;
+    const userId = req.user.id;
     const { title, content } = req.body;
     // const { userId, categoryId, title, content } = req.body;
-    await boardRepo.createPost(title, content, userId, model);
+    console.log(userId);
+    const post = await boardRepo.createPost(title, content, userId, model);
     // await freeBoardRepos.createPost(userId, categoryId, title, content);
-    return res.status(200).json({ message: "jobPosting is created" });
+    return res.status(201).json(post);
   } catch (err) {
     next(err);
   }
